refactor: 불필요한 로직 제거 및 투표 통계 서비스 리팩토링 (#367)
