+++ backend/src/test/java/com/tamnara/backend/poll/service/VoteStatisticsServiceTest.java
@@ -1,8 +1,10 @@
 package com.tamnara.backend.poll.service;
 
-import com.tamnara.backend.poll.domain.*;
-import com.tamnara.backend.poll.dto.PollStatisticsResponse;
-import com.tamnara.backend.poll.exception.PollNotFoundException;
+import com.tamnara.backend.poll.domain.Poll;
+import com.tamnara.backend.poll.domain.PollOption;
+import com.tamnara.backend.poll.domain.PollState;
+import com.tamnara.backend.poll.domain.VoteStatistics;
+import com.tamnara.backend.poll.dto.response.PollStatisticsResponse;
 import com.tamnara.backend.poll.repository.PollRepository;
 import com.tamnara.backend.poll.repository.VoteRepository;
 import com.tamnara.backend.poll.repository.VoteStatisticsRepository;
@@ -16,29 +18,28 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
 
 import static com.tamnara.backend.poll.constant.PollResponseMessage.POLL_NOT_FOUND;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class VoteStatisticsServiceTest {
 
-    @Mock
-    private PollRepository pollRepository;
+    @Mock private PollRepository pollRepository;
+    @Mock private VoteRepository voteRepository;
+    @Mock private VoteStatisticsRepository voteStatisticsRepository;
 
-    @Mock
-    private VoteRepository voteRepository;
-
-    @Mock
-    private VoteStatisticsRepository voteStatisticsRepository;
-
-    @InjectMocks
-    private VoteStatisticsService voteStatisticsService;
+    @InjectMocks private VoteStatisticsServiceImpl voteStatisticsService;
 
     private Poll poll;
     private PollOption option;
@@ -53,7 +54,7 @@ class VoteStatisticsServiceTest {
     }
 
     @Test
-    @DisplayName("generateAllStatistics 실행 시 통계가 업데이트 된다")
+    @DisplayName("generateAllStatistics 실행 시 통계 업데이트")
     void generateAllStatistics_updatesStatisticsCorrectly() {
         // given
         when(pollRepository.findByState(PollState.PUBLISHED)).thenReturn(List.of(poll));
@@ -68,7 +69,7 @@ class VoteStatisticsServiceTest {
     }
 
     @Test
-    @DisplayName("getPollStatistics는 올바른 통계 정보를 반환한다")
+    @DisplayName("getPollStatistics는 올바른 통계 정보 반환")
     void getPollStatistics_returnsCorrectResponse() {
         // given
         when(pollRepository.findById(poll.getId())).thenReturn(Optional.of(poll));
@@ -85,14 +86,18 @@ class VoteStatisticsServiceTest {
     }
 
     @Test
-    @DisplayName("getPollStatistics는 Poll이 존재하지 않을 경우 예외를 발생시킨다")
+    @DisplayName("getPollStatistics는 Poll이 존재하지 않을 경우 예외 발생")
     void getPollStatistics_throwsException_whenPollNotFound() {
         // given
         when(pollRepository.findById(1L)).thenReturn(Optional.empty());
 
-        // when & then
-        assertThatThrownBy(() -> voteStatisticsService.getPollStatistics(1L))
-                .isInstanceOf(PollNotFoundException.class)
-                .hasMessage(POLL_NOT_FOUND);
+        // when
+        ResponseStatusException exception = assertThrows(ResponseStatusException.class, () -> {
+            voteStatisticsService.getPollStatistics(1L);
+        });
+
+        // then
+        assertEquals(HttpStatus.NOT_FOUND, exception.getStatusCode());
+        assertEquals(POLL_NOT_FOUND, exception.getReason());
     }
 }
